procPhotos

procPhotos copies raw sample photos from the Fulcrum export and renames them to strain_name for use with CeNDR. The function also resizes images.

procPhotos(
  dir,
  data,
  max_dim = 500,
  overwrite = FALSE,
  CeNDR = FALSE,
 
    pub_url = "https://storage.googleapis.com/elegansvariation.org/photos/isolation/fulcrum/"
)

Arguments

dir
The path to the base fulcrum directory, raw/fulcrum/photos should contain the photos for the project

data
a data frame output from the joinGenoFulc function.

max_dim
This value sets the maximum dimension of the resized images in pixels. The default value is 500.

overwrite
Logical, passed to fs::file_copy. If TRUE then existing files with similar names will be written over. Default is FALSE.

CeNDR
Logical, determines whether to write CeNDR criteria qualifying photos to a subdirectory

Value

A folder named processed_photos in the data/processed/fulcrum directory. The folder contains full size sample photos renamed with strain names. A thumbnails subfolder is also returned within the processed_photos folder. This folder contains resized images. A dataframe identical to input data with md5 hash values and file names for all photos, also saved as an rds file in data/processed/fulcrum directory. If CeNDR is set to TRUE other subfolders will be made with those images.